requirements.txt
Flask
PyJWT
python-dotenv
werkzeug